requirements.txt
django
django-extensions
django-lifecycle
django-modeltranslation
grappelli
markdown
pymdown-extensions
pygments
git+https://github.com/vlajna95/markdown-customblocks
gunicorn
django-environ
django-on-heroku